Makefile: keep the old driver name
A bug exists in this driver that deadlocks the kernel when a net interface
managed by this driver is deleted. iwd uses the driver name
(specifically, if the driver name is "rtl88*") to figure out if it needs
to enable a workaround for this problem.

Commit fa35b99b425bc93ed712bd7c73e0701df0d46d3e changed driver name to
"88XXau" for seemingly no reason, causing this driver to be unusable for
a lot of users.

This commit changes the driver name back.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>