Bugfix: Phy cannot be moved into another namespace
When CONFIG_NET_NS is enabled one would assume the device can be
moved into another namespace. Currently it will fail with:

command failed: Operation not supported (-95)

There's a check if the kernel is between 2.6.38 and 3.0.0
before it adds the netns flag to wiphy->flags.

With this commit it will always add the netns flag when
CONFIG_NET_NS is enabled.

Signed-off-by: Daan van Gorkum <djvg@djvg.net>
Fixes: https://github.com/aircrack-ng/rtl8812au/issues/518